+++ CHANGES.rst
@@ -39,6 +39,8 @@ Release history
 - ``Model`` now has a ``connection_decode_neurons`` attribute that maps ``Connection``
   objects that require decode neurons to the corresponding ``Ensemble`` objects
   implementing them. (`#309`_)
+- Added the ``GreedyInterchip`` allocator, which reduces inter-chip communication,
+  speeding up networks with high traffic between chips. (`#309`_)
 
 **Changed**